+++ .gitignore
@@ -1,0 +1,3 @@
+/cmake-build-debug-visual-studio/
+/cmake-build-release-visual-studio/
+.idea/
