--- FunctionCallBack.cpp
@@ -1,3 +1,0 @@
-#include "FunctionCallBack.h"
-
-FunctionCallBack::FunctionCallBack() { }
