+++ include/runtime/object/FunctionCallBack.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "usings.h"
+class FunctionCallBack {
+	public:
+	FunctionCallBack();
+	virtual Object execute(List<Object> arguments) = 0;
+};
