+++ include/runtime/object/Variable.h
@@ -1,0 +1,14 @@
+#pragma once
+#include "usings.h"
+class Variable {
+	private:
+	String name;
+	String type;
+	Object value;
+	public:
+	Variable(String, String, Object);
+	String getName();
+	String getType();
+	Object getValue();
+	void setValue(Object);
+};
